Undo extraction of SparqlTripleSimple
To avoid duplication this change will only be in #1354
