+++ src/parser/CMakeLists.txt
@@ -24,7 +24,6 @@ add_library(parser
         GraphPattern.cpp data/Variable.cpp
         Iri.cpp
         Literal.cpp
-        LiteralOrIri.cpp
-        Triples.h)
+        LiteralOrIri.cpp)
 qlever_target_link_libraries(parser sparqlParser parserData sparqlExpressions rdfEscaping re2::re2 util engine)
